ENH: reorganized directories and python classes
