+++ SliceTracker/SliceTrackerUtils/steps/training.py
@@ -1,7 +1,8 @@
 import os, ast, shutil
 import qt, vtk, ctk, slicer
-from constants import SliceTrackerConstants
-from helpers import SliceTrackerStep
+
+from SliceTrackerUtils.constants import SliceTrackerConstants
+from SliceTrackerUtils.steps.base import SliceTrackerStep
 
 from SlicerProstateUtils.helpers import SampleDataDownloader
 from SlicerProstateUtils.decorators import *
